+++ packages/chai-openapi-response-validator/stryker.conf.js
@@ -15,7 +15,7 @@ module.exports = function (config) {
     transpilers: [],
     testFramework: 'mocha',
     coverageAnalysis: 'perTest',
-    thresholds: { high: 100, low: 99, break: 98 },
+    thresholds: { high: 100, low: 100, break: 100 },
     packageManager: 'yarn',
     maxConcurrentTestRunners: 1,
   });
