+++ packages/jest-openapi/src/matchers/toSatisfyApiSpec.js
@@ -4,7 +4,7 @@ const {
   EXPECTED_COLOR,
 } = require('jest-matcher-utils');
 
-const { responseFactory } = require('../openapi-validator');
+const { responseFactory } = require('openapi-validator');
 const { stringify, joinWithNewLines } = require('../utils');
 
 module.exports = function (received, openApiSpec) {
