--- packages/jest-openapi/src/openapi-validator/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib');
