--- packages/jest-openapi/src/openapi-validator/lib/classes/AbstractResponse.js
@@ -1,19 +1,0 @@
-const { stringify } = require('../utils');
-
-class AbstractResponse {
-  constructor(res) {
-    this.res = res;
-  }
-
-  summary() {
-    return {
-      body: this.body,
-    };
-  }
-
-  toString() {
-    return stringify(this.summary());
-  }
-}
-
-module.exports = AbstractResponse;
