--- packages/jest-openapi/src/openapi-validator/lib/classes/RequestPromiseResponse.js
@@ -1,29 +1,0 @@
-const AbstractResponse = require('./AbstractResponse');
-
-class RequestPromiseResponse extends AbstractResponse {
-  constructor(res) {
-    super(res);
-    this.status = res.statusCode;
-    this.body = res.request._json // eslint-disable-line no-underscore-dangle
-      ? res.body
-      : res.body.replace(/"/g, '');
-    this.req = res.req;
-    this.bodyHasNoContent = this.body === '';
-  }
-
-  getBodyForValidation() {
-    if (this.bodyHasNoContent) {
-      return null;
-    }
-    try {
-      return JSON.parse(this.body);
-    } catch (error) {
-      // if JSON.parse errors, then body is not stringfied JSON that
-      // needs parsing into a JSON object, so just move to the next
-      // block and return the body
-    }
-    return this.body;
-  }
-}
-
-module.exports = RequestPromiseResponse;
