--- packages/jest-openapi/src/openapi-validator/lib/classes/errors/ValidationError.js
@@ -1,12 +1,0 @@
-class ValidationError extends Error {
-  constructor(code, message) {
-    super(message);
-    this.code = code;
-  }
-
-  toString() {
-    return this.message;
-  }
-}
-
-module.exports = ValidationError;
