--- packages/jest-openapi/src/openapi-validator/lib/index.js
@@ -1,7 +1,0 @@
-const openApiSpecFactory = require('./openApiSpecFactory');
-const responseFactory = require('./responseFactory');
-
-module.exports = {
-  openApiSpecFactory,
-  responseFactory,
-};
