--- packages/jest-openapi/src/openapi-validator/package.json
@@ -1,17 +1,0 @@
-{
-  "private": true,
-  "scripts": {},
-  "devDependencies": {
-    "chai": "^4.2.0",
-    "eslint": "^6.8.0"
-  },
-  "dependencies": {
-    "compress-tag": "^2.0.0",
-    "fs-extra": "^9.0.0",
-    "js-yaml": "^3.13.1",
-    "openapi-response-validator": "^4.0.0",
-    "openapi-schema-validator": "^3.0.3",
-    "path-parser": "^6.1.0",
-    "typeof": "^1.0.0"
-  }
-}
