+++ packages/openapi-validator/README.md
@@ -1,0 +1,6 @@
+# OpenAPI Validator
+
+[![downloads](https://img.shields.io/npm/dm/openapi-validator)](https://www.npmjs.com/package/openapi-validator)
+[![npm](https://img.shields.io/npm/v/openapi-validator.svg)](https://www.npmjs.com/package/openapi-validator)
+
+Common code for [jest-openapi](https://www.npmjs.com/package/jest-openapi) and [Chai OpenAPI Response Validator](https://www.npmjs.com/package/chai-openapi-response-validator)
