+++ packages/openapi-validator/lib/classes/AbstractOpenApiSpec.js
