+++ packages/openapi-validator/lib/classes/AbstractResponse.js
