+++ packages/openapi-validator/lib/classes/OpenApi2Spec.js
