+++ packages/openapi-validator/lib/classes/OpenApi3Spec.js
