+++ packages/openapi-validator/lib/classes/errors/ValidationError.js
