+++ packages/openapi-validator/lib/openApiSpecFactory.js
