+++ packages/openapi-validator/lib/utils.js
